- Guarantee that the 'LocalFeedLoader' does not deliver deletion error after instance has been deallocated - Moved 'LocalFeedLoader' class and 'FeedStore' protocol to the production the - 'LocalFeedLoader' class and 'FeedStore' protocol has been made public
