+++ EssentialFeed_2023/EssentialFeed_2023/FeedCache/FeedStore.swift
@@ -1,0 +1,17 @@
+//
+//  FeedStore.swift
+//  EssentialFeed_2023
+//
+//  Created by Fenominall on 5/3/23.
+//
+
+import Foundation
+
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+    
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
